Move Reflection Utils into Tests folder Fixes #25
